Increase import performance for delimited-ascii files by using CStdioFile::ReadString instead of File::ReadLnAscii
git-svn-id: https://svn.52north.org/svn/ilwis/drawers@2540 b64c5c0c-1e74-440b-9268-4ad09cf26390

+++ Sources/Engine/DataExchange/TableDBF.cpp
@@ -436,7 +436,7 @@ void TableDBF::AddToTableDomain(FieldValues &line, int iKeyColumn)
 		(*cur).iValues.push_back(iUNDEF);
 }
 
-int TableDBF::iNewRecord(int iRec)
+long TableDBF::iNewRecord(long iRec)
 {
 	for(vector<MemoVector>::iterator cur = memos.begin(); cur != memos.end(); ++cur)
 		(*cur).iValues.push_back(iUNDEF);
